+++ Runtime/Unity.Serialization/Json/Adapters/JsonUnityObjectAdapter.cs
@@ -4,6 +4,9 @@ namespace Unity.Serialization.Json
 {
     internal class JsonUnityObjectAdapter : JsonVisitorAdapter
         , IVisitAdapter<UnityEngine.Object>
+#if UNITY_EDITOR
+        , IVisitAdapter<UnityEditor.GlobalObjectId>
+#endif
         , IVisitContainerAdapter
     {
         public JsonUnityObjectAdapter(JsonVisitor visitor) : base(visitor) { }
@@ -17,6 +20,16 @@ namespace Unity.Serialization.Json
 #endif
             return VisitStatus.Override;
         }
+
+#if UNITY_EDITOR
+        public VisitStatus Visit<TProperty, TContainer>(IPropertyVisitor visitor, TProperty property, ref TContainer container, ref UnityEditor.GlobalObjectId value, ref ChangeTracker changeTracker)
+            where TProperty : IProperty<TContainer, UnityEditor.GlobalObjectId>
+        {
+            var str = EncodeJsonString(value.ToString());
+            Append(property, str, (builder, s) => { builder.Append(s); });
+            return VisitStatus.Override;
+        }
+#endif
 
         public VisitStatus BeginContainer<TProperty, TValue, TContainer>(IPropertyVisitor visitor, TProperty property,
             ref TContainer container, ref TValue value, ref ChangeTracker changeTracker) where TProperty : IProperty<TContainer, TValue>