+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.1-preview] - 2019-10-21
+### Added
+* Support JSON serialization of `UnityEditor.GlobalObjectId`.
+* Support JSON serialization of `UnityEditor.GUID`.
+* New method `DeserializeFromStream` to deserialize from stream object.
+
+### Changed
+* Updated `com.unity.properties` to version `0.9.1-preview`.
+* Deserialization will now attempt to construct the destination container using `PropertyContainer.Construct` utility.
+* Deserialization will now attempt to read type information field `$type` by default.
+
 ## [0.5.0-preview] - 2019-10-07
 ### Changed
 * Updated `com.unity.properties` to version `0.9.0-preview`.